pyproject.toml
[tool.poetry]
name = "rpc"
version = "0.1.0"
description = ""
authors = ["BigBlackBob-93 <101857203+BigBlackBob-93@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
grpcio-tools = "^1.59.3"
grpcio = "^1.59.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
